refactor: create pattern-units folder
