+++ src/main.tsx
@@ -6,7 +6,7 @@ import { Provider } from 'react-redux';
 import { store } from './store/store.ts';
 import { QueryClientProvider } from '@tanstack/react-query';
 import { queryClient } from './lib/react-query.ts';
-import './i18n.ts';
+import './pattern-units/i18n.ts';
 import { BrowserRouter } from 'react-router-dom';
 
 export function renderApp(container: HTMLElement) {
