+++ src/pattern-units/i18n.ts
@@ -2,8 +2,8 @@
 import i18n from 'i18next';
 import { initReactI18next } from 'react-i18next';
 
-import enTranslation from './locales/en.json';
-import ruTranslation from './locales/ru.json';
+import enTranslation from '../locales/en.json';
+import ruTranslation from '../locales/ru.json';
 
 void i18n
   .use(initReactI18next) // Passes i18next to react-i18next
