+++ src/tests/test1.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { describe, it, expect } from 'vitest';
-import Hello from '../hello';
+import Hello from '../pattern-units/hello.tsx';
 
 describe('Hello component', () => {
   it('renders greeting', () => {
